fix: styling
I wasn't a fan of the "enter details manually" item having the same margin as a server listing to make up for the status indicator.

It still has the same margin on the <ul> just to keep things aligned, but it DOES NOT have the margin inside the button itself.

Idk.. just looks better to me.